# Simplify `Permutations`