Fixes #747

This is quite a rewrite, very little is untouched.
`count` and `size_hint` are really simplified with the new `PermutationState::size_hint_for`.

    cargo bench permutation
    // Before this PR
    permutations iter       [62.130 µs 62.355 µs 62.620 µs]
    permutations range      [61.637 µs 61.850 µs 62.104 µs]
    permutations slice      [64.177 µs 64.314 µs 64.459 µs]
    // Now
    permutations iter       [57.280 µs 57.478 µs 57.702 µs]    -8.4%
    permutations range      [59.714 µs 59.914 µs 60.138 µs]    -4.7%
    permutations slice      [59.584 µs 59.720 µs 59.868 µs]    -7.1%
I did not expect any performance improvement, but nice to have a little one.