@Philippe-Cholet 2023-11-01
I'm sorry you had some hard time reviewing it. I definitely tried and struggled to make more fine-grained commits here. But I should have thought of your very helpful intermediary `Loaded(CompleteState) -> LoadedStart, LoadedOngoing`.
I'm a work in progress (especially git), and to improve myself I splitted it again myself, trying to leave the last commit intact (`FusedIterator`, but I git-failed so I had to write it again).

You can check that there is no difference between our branches.