@Philippe-Cholet 2023-11-01
@phimuemue I did not see any offense, only the opportunity to improve the way I write commits.